+++ CHANGELOG.md
@@ -1,15 +1,18 @@
 # Changelog
 
-## [0.1.2] - 2023-10-10
+## [0.1.2] - 2023-10-30
+- Create SparkScopeApp with CLI to run SparkScope as standalone app 
+  - reads application context, spark conf and events from eventLog  
+  - can be run on finished or running spark application
 - Rename package from sparkscope to sparkscope-spark2
 - Bump Scala version from 2.11.8 to 2.11.12
 - Handle gently case for master=local where executorId=driver
 
 ## [0.1.1] - 2023-10-06
 - configuration:
-  - sparkscope will prioritize csv metrics SparkConf configuration and use metrics.properties files as last-resort 
+  - sparkscope will prioritize csv metrics SparkConf configuration and use metrics.properties files as last-resort
 - html-report:
-    - remove core-hours and heap gb-hours from warnings
+  - remove core-hours and heap gb-hours from warnings
 
 ## [0.1.0] - 2023-09-29
 Initial SparkScope release: